tests/requirements-dev.txt
pypdf==5.1.0
azure-identity==1.16.1
azure-search-documents==11.4.0b11
azure-storage-blob==12.20.0
openai==1.35.8
rich==13.9.4